+++ hypervisor/Makefile
@@ -189,7 +189,6 @@ C_SRCS += arch/x86/guest/vmexit.c
 S_SRCS += arch/x86/guest/vmx_asm.S
 C_SRCS += arch/x86/guest/trusty.c
 C_SRCS += arch/x86/cat.c
-C_SRCS += lib/misc.c
 C_SRCS += lib/string.c
 C_SRCS += lib/memory.c
 C_SRCS += lib/crypto/crypto_api.c
